requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.25.0
fastapi>=0.109.0
uvicorn>=0.27.0
aiofiles>=23.2.0
jinja2>=3.1.0
python-multipart>=0.0.6
